PORTALS-2582
- decouple SynapseContext from QueryClientContext and ThemeContext
- Add FullSynapseContext to provide backwards-compatibility

SageAccountWeb handles its own MUI theme and QueryClient. It makes more sense to allow SageAccountWeb to just pull in the SynapseContext that it needs instead of having multiple (potentially-conflicting) MUI themes and QueryClient contexts. This commit makes that change possible.
