+++ apps/SageAccountWeb/src/AppInitializer.tsx
@@ -3,7 +3,7 @@ import React, { useCallback, useEffect, useState } from 'react'
 import { AppContextProvider } from 'AppContext'
 import { Redirect, useLocation } from 'react-router-dom'
 import { SynapseClient, SynapseComponents } from 'synapse-react-client'
-import { SynapseContextProvider } from 'synapse-react-client/dist/utils/SynapseContext'
+import { FullContextProvider as SynapseContextProvider } from 'synapse-react-client/dist/utils/FullContextProvider'
 import { UserProfile } from 'synapse-react-client/dist/utils/synapseTypes'
 import { getSearchParam } from 'URLUtils'
 import useAnalytics from './useAnalytics'
