+++ packages/synapse-react-client/package.json
@@ -227,7 +227,8 @@
     "link-src": "pnpm link && cd ./node_modules/react && pnpm link && cd ../react-router && pnpm link && cd ../react-router-dom && pnpm link && cd ../../",
     "unlink-src": "pnpm unlink && cd ./node_modules/react && pnpm unlink && cd ../react-router && pnpm unlink && cd ../react-router-dom && pnpm unlink && cd ../../",
     "storybook": "storybook dev -p 6060",
-    "lint": "eslint src"
+    "lint": "eslint src",
+    "watch:js": "tsc --watch"
   },
   "resolutions": {
     "js-yaml": "3.13.1",
