+++ packages/synapse-react-client/src/lib/containers/StorybookComponentWrapper.tsx
@@ -1,10 +1,6 @@
 import React, { useEffect, useMemo } from 'react'
 import { MemoryRouter } from 'react-router-dom'
-import {
-  defaultQueryClientConfig,
-  SynapseContextProvider,
-  SynapseContextType,
-} from '../utils/SynapseContext'
+import { SynapseContextType } from '../utils/SynapseContext'
 import { QueryClient } from 'react-query'
 import { SynapseClient } from '../utils'
 import { SynapseToastContainer } from './ToastMessage'
@@ -12,8 +8,8 @@ import { ReactQueryDevtools } from 'react-query/devtools'
 import dayjs from 'dayjs'
 import {
   getAccessTokenFromCookie,
-  getUserProfile,
   getAuthenticatedOn,
+  getUserProfile,
   signOut,
 } from '../utils/SynapseClient'
 import { SynapseClientError } from '../utils/SynapseClientError'
@@ -34,6 +30,10 @@ import {
   stopAdPortalPalette,
 } from '../utils/theme/palette/Palettes'
 import useDetectSSOCode from '../utils/hooks/useDetectSSOCode'
+import {
+  defaultQueryClientConfig,
+  FullContextProvider,
+} from '../utils/FullContextProvider'
 
 export async function sessionChangeHandler() {
   let accessToken: string | undefined = await getAccessTokenFromCookie()
@@ -135,7 +135,7 @@ export function StorybookComponentWrapper(props: {
   )
 
   return (
-    <SynapseContextProvider
+    <FullContextProvider
       queryClient={storybookQueryClient}
       key={accessToken}
       synapseContext={synapseContext}
@@ -151,7 +151,7 @@ export function StorybookComponentWrapper(props: {
         <SynapseToastContainer />
         <main>{props.children}</main>
       </MemoryRouter>
-    </SynapseContextProvider>
+    </FullContextProvider>
   )
 }
