+++ packages/synapse-react-client/src/lib/utils/theme/useTheme.tsx
@@ -3,8 +3,11 @@ import { ThemeProvider as MuiThemeProvider } from '@mui/material/styles'
 import { merge } from 'lodash-es'
 import React, { useMemo } from 'react'
 import defaultMuiTheme from './DefaultTheme'
+import type { PartialDeep } from 'type-fest'
 
-export function mergeTheme(themeOverrides: ThemeOptions): ThemeOptions {
+export function mergeTheme(
+  themeOverrides: PartialDeep<ThemeOptions>,
+): ThemeOptions {
   // TODO: Handle merging color palettes where an entire palette can be generated from a single base color.
   return merge({}, defaultMuiTheme, themeOverrides)
 }
