+++ packages/synapse-react-client/test/lib/utils/hooks/SynapseAPI/useEntityBundle.test.tsx
@@ -4,17 +4,17 @@ import { QueryClient } from 'react-query'
 import useGetEntityBundle from '../../../../../src/lib/utils/hooks/SynapseAPI/entity/useEntityBundle'
 import { EntityBundle } from '../../../../../src/lib/utils/synapseTypes'
 import { MOCK_CONTEXT_VALUE } from '../../../../../mocks/MockSynapseContext'
-import { SynapseContextProvider } from '../../../../../src/lib/utils/SynapseContext'
+import FullContextProvider from '../../../../../src/lib/utils/FullContextProvider'
 
 const queryClient = new QueryClient()
 
 const wrapper = (props: { children: React.ReactChildren }) => (
-  <SynapseContextProvider
+  <FullContextProvider
     synapseContext={MOCK_CONTEXT_VALUE}
     queryClient={queryClient}
   >
     {props.children}
-  </SynapseContextProvider>
+  </FullContextProvider>
 )
 
 const expected: EntityBundle = {
