+++ packages/synapse-react-client/test/lib/utils/hooks/SynapseAPI/useProjects.test.tsx
@@ -12,16 +12,17 @@ import {
 import { MOCK_CONTEXT_VALUE } from '../../../../../mocks/MockSynapseContext'
 import { QueryClient } from 'react-query'
 import { SynapseContextProvider } from '../../../../../src/lib/utils/SynapseContext'
+import FullContextProvider from '../../../../../src/lib/utils/FullContextProvider'
 
 const queryClient = new QueryClient()
 
 const wrapper = (props: { children: React.ReactChildren }) => (
-  <SynapseContextProvider
+  <FullContextProvider
     synapseContext={MOCK_CONTEXT_VALUE}
     queryClient={queryClient}
   >
     {props.children}
-  </SynapseContextProvider>
+  </FullContextProvider>
 )
 
 const request: EntityChildrenRequest = {
